Movement Syncing Works! [see details]
It looks like nothing's happening, but there were a lot of days of implementing and testing and then deleting and restructuring...

Basically multiple headsets were going to sync their locations, but I found an issue where the 2nd joined headset player won't have display (multiple OVRCameraRig in the same scene was fighting each other, but I wasn't aware)

So, new solution was to have a local OVRCR and sync their location to a network representation with RPC calls. Super easy to setup and worked without issue.