added scenic pause
Pause implemented in TimelineManager. ZMQ will call "NotifyScenicPause" on TimelineManager on every frame when new scenic data for "pause" comes in. pause boolean is a global variable stored on ego that can be called by any scenic agent to stop the whole scene on unity side, except the ego. currently, it does not affect the running status of scenic scene.